Ajout d'un fonctionnalité pour selectioner un user pour un note en tant que admin

+++ views/admin/admin.view.php
@@ -1,8 +1,7 @@
 <?php require 'views/partials/header.php'; ?>
 
-
 <h1>Dashboard</h1>
-<button><a href="/note-new">Crée une nouvelle note</a></button>
+<button><a href="/note-new">Créer une nouvelle note</a></button>
 <table border="1">
     <tr>
         <th>Note ID</th>
@@ -16,36 +15,38 @@
 
     <?php if (!empty($notes)) : ?>
         <?php foreach ($notes as $note) : ?>
-            <tr>
-                <td><?= $note['Note_ID']; ?></td>
-                <td><?= $note['User_ID']; ?></td>
-                <td><?= $note['Title']; ?></td>
-                <td><?= $note['Content']; ?></td>
-                <td>
-                    <?php
-                    $imageName = $note['File_Name'];
-                    if (!empty($imageName)) {
-                        $imagePath = "/uploads/$imageName";
-                        echo "<img src='$imagePath' alt='Image' style='max-width: 80px; max-height: 80px;'>";
-                    } else {
-                        echo "Aucune image";
-                    }
-                    ?>
-                </td>
+            <?php if ($note['Is_Admin'] != 1) : ?>
+                <tr>
+                    <td><?= $note['Note_ID']; ?></td>
+                    <td><?= $note['User_ID']; ?></td>
+                    <td><?= $note['Title']; ?></td>
+                    <td><?= $note['Content']; ?></td>
+                    <td>
+                        <?php
+                        $imageName = $note['File_Name'];
+                        if (!empty($imageName)) {
+                            $imagePath = "/uploads/$imageName";
+                            echo "<img src='$imagePath' alt='Image' style='max-width: 80px; max-height: 80px;'>";
+                        } else {
+                            echo "Aucune image";
+                        }
+                        ?>
+                    </td>
 
-                <td><?= $note['Name']; ?></td>
-                <td class="actions">
-                    <?php
-                    $noteId = $note['Note_ID'];
-                    $viewLink = "note?id=$noteId";
-                    $updateLink = "note-update?id=$noteId";
-                    $deleteLink = "note-delete?id=$noteId";
-                    ?>
-                    <button><a href="<?= $viewLink ?>">Voir</a></button>
-                    <button><a href="<?= $updateLink ?>">Modifier</a></button>
-                    <button><a href="<?= $deleteLink ?>" onclick="return confirm('Êtes-vous sûr de vouloir supprimer cette note ?')">Supprimer</a></button>
-                </td>
-            </tr>
+                    <td><?= $note['Name']; ?></td>
+                    <td class="actions">
+                        <?php
+                        $noteId = $note['Note_ID'];
+                        $viewLink = "note?id=$noteId";
+                        $updateLink = "note-update?id=$noteId";
+                        $deleteLink = "note-delete?id=$noteId";
+                        ?>
+                        <button><a href="<?= $viewLink ?>">Voir</a></button>
+                        <button><a href="<?= $updateLink ?>">Modifier</a></button>
+                        <button><a href="<?= $deleteLink ?>" onclick="return confirm('Êtes-vous sûr de vouloir supprimer cette note ?')">Supprimer</a></button>
+                    </td>
+                </tr>
+            <?php endif; ?>
         <?php endforeach; ?>
     <?php else : ?>
         <tr>
